api: merge os module into main api reference
Background: After using this for a while, I think moving this
out into a separate page was premature. You really want one page that you can
use to ctrl+f for functions.

I also wonder if we should host our own copy of the starlark built-in methods
(like strings and lists and dicts)