docs: add subnav with top reference pages (#888)
We've got a LOT of docs, but most of the time when folks end up
on the docs site, they probably want the API or CLI reference,
which are waaaay at the bottom.

This adds a subnav that includes a few links to the (hopefully)
most important pages. The search comes along for a ride out of
the sidebar as well.

There were some...interesting things going on with the layout,
so there's some general changes here to flexbox more flexbox-y,
but they should not actually change things, just make it more
friendly for future adaptions (like this one).