docs: make entire header a permalink
This behavior mimics how GitHub (and many other sites) behave,
which is a pretty familiar pattern for users.

The entire header is a permalink now, and the '🔗' shows up on
hover (replacing `►`).

(As a sidenote, the permalink functionality does not really need
Javascript - we should arguably rewrite this as a Jekyll include
to set up the permalinks at generation time, but at the same time,
functionality without it isn't hugely degraded, so it's not worth
it right now.)