docs: sticky sidebar nav (#907)
On scroll, the sidebar nav will stay sticky at the top of the
screen.

You can still scroll to the bottom of the content area (i.e. the
footer links) and it _will_ scroll out at that point. This behavior
is pretty consistent with other sticky navs on a few docs sites I
played with.