docs: copy to clipboard + code styling tweaks
Big addition here is a "Copy" button next to each code block to
copy to clipboard (uses `clipboard.js` - works in all modern
browsers and does not use Flash).

As part of this, the code styling was tweaked slightly to improve
spacing: the code blocks are closer to the element that precedes
them, which is logically how the docs flow/intend them to be used.
(For this reason, the button is on the bottom!)

The "Install" page also had its Markdown tweaked to fix the way
the list elements were organized to prevent breaking a list into
multiple, which is both semantically better and helps the spacing
on the content substantially.