docs: re-enable links opening in external window
Re-add the `jekyll-target-blank` plugin but adjust the Google Custom
Search so that it doesn't break it.

The search fix is to use a `<div>` instead of the custom
`<gcse:searchresults-only>` element (which was erroneously having its
`gcse` "namespace" chopped off by nokogiri during DOM processing). This
is actually the way that Google has it in their docs now as of Aug 2021
and works fine.

While it's not great that nokogiri can't handle this properly, the
likelihood we use custom elements in the future with prefixed namespaces
is minimal, and the alternatives are:
 * Use JS to add `target="_blank"` to all links on page load
 * Remember to specify it manually on all external links