docs: more specific help pointers
In technical documentation, I don't think it's helpful to be constantly saying
"please suggest changes to this API!"

Most readers are just trying to understand how to get their app working --
and will not be well-equipped to suggest changes and it's weird to ask them to.

The only exception is when we have specific pointers to issues
or documentation on how to script custom behavior.