docs: small corrections to probe() arguments (#963)
Fix the default value for `timeout` to be correct (1sec). This
has always been the case since introduction, docs were just wrong.

Small typo fix as well, and the default values are reiterated in
the argument descriptions since they're easy to miss in the function
signature but are very important.