blog: revamp "Switch from Docker Desktop to Rancher Desktop" (#1067)
Update based on ecosystem changes since originally published in
Sept 2021. (If you're keeping score, Nick predicted this right
away.)

The original post, which is focused on the now inactive `kim`
builder experiment that shipped with early versions of Rancher
Desktop is now moved to a "new" (but backdated) post for posterity,
as it's not a bad example of writing a `custom_build` extension.
It also links to the "Switch to Rancher Desktop" post for any
Rancher Desktop users that stumble upon it, as it's now no longer
relevant for them.

The new post is pretty short and sweet because there's not a ton
to do! But it lives at the old slug (SEO yay) and _really_ fulfills
its promise of a 5 minute swap :)

See also tilt-dev/tilt#5513, which adds the in-cluster detection
support, so is required for this to work as expected.